refactor(skill): refactor req body validation

+++ src/app/api/skills/[id]/route.ts
@@ -1,10 +1,18 @@
+import { authOptions } from '@/lib/prisma';
 import { skillService } from '@/skill/SkillService';
-import { HTTPError, UnprocessableEntityError } from '@/utils/error';
+import { BadRequestError, HTTPError, UnauthenticatedError, UnprocessableEntityError } from '@/utils/error';
 
+import { getServerSession } from 'next-auth';
 import { NextResponse } from 'next/server';
 import { z } from 'zod';
 
 const updateSkillSchema = z.object({
+  id: z
+    .string({
+      required_error: 'Id is required',
+      invalid_type_error: 'id must be a string',
+    })
+    .cuid('Id is invalid'),
   name: z
     .string({
       invalid_type_error: 'Name must be a string',
@@ -25,20 +33,31 @@ export const PUT = async (req: Request, {
   params: { id: string };
 }) => {
   const id = params.id;
+
   try {
+    const session = await getServerSession(authOptions);
+    if (!session) {
+      throw new UnauthenticatedError('You should login first to access this endpoint');
+    }
+
     const formData = await req.formData();
 
-    const name = formData.get('name') as string | undefined;
-    const categoryId = formData.get('category') as string | undefined;
-
-    const result = updateSkillSchema.safeParse({ name, categoryId });
+    const result = updateSkillSchema.safeParse({
+      id,
+      name: formData.get('name'),
+      categoryId: formData.get('category'),
+    });
     if (!result.success) {
       throw new UnprocessableEntityError(result.error.issues[0].message);
     }
 
     const photo = formData.get('photo') as File | undefined;
 
-    const response = await skillService.updateSkill(id, { name, categoryId, photo });
+    const response = await skillService.updateSkill(result.data.id, {
+      name: result.data.name,
+      categoryId: result.data.categoryId,
+      photo,
+    });
 
     return NextResponse.json(response);
   } catch (error) {
@@ -52,6 +71,15 @@ export const PUT = async (req: Request, {
   }
 };
 
+const deleteSkillSchema = z.object({
+  id: z
+    .string({
+      required_error: 'Id is required',
+      invalid_type_error: 'id must be a string',
+    })
+    .cuid('Id is invalid'),
+});
+
 export const DELETE = async (_: Request, {
   params,
 }: {
@@ -60,6 +88,16 @@ export const DELETE = async (_: Request, {
   const id = params.id;
 
   try {
+    const session = await getServerSession(authOptions);
+    if (!session) {
+      throw new UnauthenticatedError('You should login first to access this endpoint');
+    }
+
+    const result = deleteSkillSchema.safeParse({ id });
+    if (!result.success) {
+      throw new BadRequestError(result.error.issues[0].message);
+    }
+
     const response = await skillService.deleteSkill(id);
 
     return NextResponse.json(response);
